Add dynamic image and crop size input for mosaic generator
Updated main.py to prompt the user for the image filename and crop size, allowing mosaic parts of any size. Changed hardcoded directory and crop size to variables, and updated the criteria file to reflect that mosaic parts can be any size but will be resized to a square format.